+++ src/Domain/Data/Entities/Media.cs
@@ -7,9 +7,9 @@ namespace Mublog.Server.Domain.Data.Entities
     {
         public Guid PublicId { get; set; }
         public MediaType MediaType { get; set; }
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
         public Profile Owner { get; set; }
-        public int PostId { get; set; }
+        public long PostId { get; set; }
         public Post ParentPost { get; set; }
     }
 }
